+++ src/js/language/be_FR.js
@@ -237,6 +237,7 @@
                 countryNotSupported: 'Le code de pays %s n\'est pas accepté',
                 country: 'Veuillez fournir un numéro de téléphone valide pour %s',
                 countries: {
+                    AE: 'Émirats Arabes Unis',
                     BR: 'Brésil',
                     CN: 'Chine',
                     CZ: 'Tchèque',