+++ src/js/language/es_ES.js
@@ -237,6 +237,7 @@
                 countryNotSupported: 'El código del país %s no está soportado',
                 country: 'Por favor introduce un número válido de teléfono en %s',
                 countries: {
+                    AE: 'Emiratos Árabes Unidos',
                     BR: 'Brasil',
                     CN: 'China',
                     CZ: 'República Checa',